Added commit functions within Helper file
Added commit function within the Helper file to ensure compliance with automated tester requirements.

Data from auto populator needed to be sequenced in a particular sequence, hence the addition of the commit function within the Helper file.

This ensures that during the populate execution period, the data being populated is sequenced in the way the tester expects this.